+++ src/LSST/system/ComServer.cpp
@@ -27,6 +27,7 @@
 #include <thread>
 
 // Project headers
+#include "system/ComClient.h"
 #include "system/Config.h"
 #include "util/Log.h"
 
@@ -52,7 +53,25 @@ ComServer::ComServer(IoContextPtr const& ioContext, int port)
     _acceptor.set_option(boost::asio::socket_base::reuse_address(true));
 }
 
-ComServer::~ComServer() { LDEBUG("ComServer::~ComServer()"); }
+ComServer::~ComServer() {
+    LDEBUG("ComServer::~ComServer()");
+    if (!_shutdown) {
+        LERROR("ComServer::~ComServer() shudown wasn't called, calling now");
+        shutdown();
+    }
+    // Start a client connection to this server to get the
+    // accept loop to move. Catch and ignore any problems
+    // as the `cmd` is expected to fail.
+    try {
+        // The server must be on this host.
+        ComClient client(_ioContext, "127.0.0.1", _port);
+        string cmd("ComServer destructor shuting down");
+        client.writeCommand(cmd);
+        LDEBUG("wrote cmd=", cmd);
+    } catch (std::exception const& ex) {
+        LWARN("Ignoring shutdown issue ", ex.what());
+    }
+}
 
 void ComServer::run() {
     LDEBUG("ComServer::run()");