+++ src/LSST/util/Log.h
@@ -141,6 +141,18 @@ public:
     /// BUFFER or MIRRORED as the destination.
     void setMaxBuffers(int maxBuffers);
 
+    /// @return value of environment variabe LOGLVL, defaults to TRACE (1).
+    /// Expected values of LOGLVL are associated with their corresponding LogLvl.
+    /// TRACE being 1, CRITICAL being 6. Values are constrained such that
+    /// values less than TRACE are set to TRACE, and values greater than
+    /// CRITCAL are set to CRITICAL.
+    static LogLvl getEnvironmentLogLvl();
+
+    /// Set the current _logLvl to the environment LOGLVL
+    /// Useful for quieting down local unit tests.
+    /// @see Log::getEnvironmentLogLvl()
+    void useEnvironmentLogLvl();
+
 private:
     /// Private constructor as only instance of Log should be availabel through getLog().
     /// Log needs to be available before the configuration is read so that configuration